include libraries and configure hilt
