+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,10 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.marvelcharacters">
 
+    <uses-permission android:name="android.permission.INTERNET" />
+
     <application
+        android:name=".MarvelCharactersApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -10,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MarvelCharacters">
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.marvelcharacters.presentation.ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MarvelCharacters">
