+++ app/src/main/java/com/example/marvelcharacters/Application.kt
@@ -1,0 +1,7 @@
+package com.example.marvelcharacters
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MarvelCharactersApp: Application()
