+++ app/src/main/java/com/example/marvelcharacters/presentation/ui/MainActivity.kt
@@ -1,0 +1,19 @@
+package com.example.marvelcharacters.presentation.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import com.example.marvelcharacters.presentation.ui.theme.MarvelCharactersTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            MarvelCharactersTheme {
+
+            }
+        }
+    }
+}
