+++ app/src/main/java/com/example/marvelcharacters/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.marvelcharacters.ui.theme
+package com.example.marvelcharacters.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
