+++ build.gradle
@@ -1,11 +1,14 @@
 buildscript {
     ext {
-        compose_version = '1.1.0'
+        compose_version = '1.1.1'
+    }
+    dependencies {
+        classpath "com.google.dagger:hilt-android-gradle-plugin:2.40.5"
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id 'com.android.application' version '7.1.3' apply false
-    id 'com.android.library' version '7.1.3' apply false
+    id 'com.android.application' version '7.1.0' apply false
+    id 'com.android.library' version '7.1.0' apply false
     id 'org.jetbrains.kotlin.android' version '1.6.10' apply false
 }
